pyproject.toml
[tool.poetry]
name = "pcd-sp-2024"
version = "0.1.0"
description = "Mi primer proyecto cd"
authors = ["izainea <cizaineam@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
scikit-learn = "^1.5.0"
pandas = "^2.2.2"
numpy = "^1.26.4"
matplotlib = "^3.9.0"
seaborn = "^0.13.2"
ucimlrepo = "^0.0.7"
sodapy = "^2.2.0"
python-dotenv = "^1.0.1"
streamlit = "^1.36.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
